+++ components/layout/Sidebar.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { SidebarHeader } from '../sidebar/SidebarHeader'
 import { SidebarSearch } from '../sidebar/SidebarSearch'
 import { SidebarFooter } from '../sidebar/SidebarFooter'
